my agent test

+++ python/gps/agent/box2d/my_agent_box2d.py
@@ -9,7 +9,7 @@ from gps.sample.sample import Sample
 from gps.proto.gps_pb2 import JOINT_ANGLES
 
 import rospy
-from std_msgs.msg import Float64
+from std_msgs.msg import Float32
 
 class AgentBox2D(Agent):
     """
@@ -22,8 +22,8 @@ class AgentBox2D(Agent):
         Agent.__init__(self, config)
 
         rospy.init_node('dummy_agent', anonymous=True)
-        self.pub = rospy.Publisher('des_dis', Float64, queue_size=1)
-        self.sub = rospy.Subscriber('des_dis', Float64, self.callback)
+        self.pub = rospy.Publisher('des_dis', Float32, queue_size=1)
+        self.sub = rospy.Subscriber('des_dis', Float32, self.callback)
 
         self.x0 = self._hyperparams["x0"]
